pyproject.toml
[project]
name = "speak-bot"
version = "0.1.0"
description = "A voice-first assistant pipeline."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "openai-whisper",
    "scipy>=1.10.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "piper-tts>=1.2.0",
]

[tool.setuptools]
packages = ["speak_bot"]
